corrections after PR review
