+++ app/screens/post_options/components/options/copy_text_option.tsx
@@ -4,6 +4,7 @@
 import Clipboard from '@react-native-community/clipboard';
 import React, {useCallback} from 'react';
 
+import {Screens} from '@constants';
 import {t} from '@i18n';
 import {dismissBottomSheet} from '@screens/navigation';
 
@@ -15,7 +16,7 @@ type Props = {
 const CopyTextOption = ({postMessage}: Props) => {
     const handleCopyText = useCallback(() => {
         Clipboard.setString(postMessage);
-        dismissBottomSheet();
+        dismissBottomSheet(Screens.POST_OPTIONS);
     }, [postMessage]);
 
     return (
